feat: finish 2565.cpp (the type of the data is mistaken, 要求保留三位小数, to fix the bug, I change 'int' into 'float')

+++ xmx_cpp/2565.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 double Eval_Polynomial(const double a[], int n, double x){
-    int count = 0;
+    double count = 0.000;
     for (int i = 0; i < n + 1; i++){
         count += a[i] * pow(x, i);
     }
